Changed float32 to float64 and changed a few topics

+++ captain_interface/include/captain_interface/RosInterFace/RosInterFace.h
@@ -12,7 +12,7 @@
 
 //Ros messages
 #include <std_msgs/String.h>
-#include <std_msgs/Float32.h>
+#include <std_msgs/Float64.h>
 #include <std_msgs/Empty.h>
 #include <std_msgs/Header.h>
 #include <geometry_msgs/PoseWithCovarianceStamped.h>
@@ -135,17 +135,17 @@ struct RosInterFace {
   void ros_callback_done(const std_msgs::Empty::ConstPtr &_msg);
   //void ros_callback_waypoint(const smarc_msgs::LatLonStamped::ConstPtr &_msg);
   void ros_callback_waypoint(const geographic_msgs::GeoPoint::ConstPtr &_msg);
-  void ros_callback_speed(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_depth(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_altitude(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_yaw(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_yawrate(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_pitch(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_rpm(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_rudder(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_elevator(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_thrusterPort(const std_msgs::Float32::ConstPtr &_msg);
-  void ros_callback_thrusterStrb(const std_msgs::Float32::ConstPtr &_msg);
+  void ros_callback_speed(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_depth(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_altitude(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_yaw(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_yawrate(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_pitch(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_rpm(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_rudder(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_elevator(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_thrusterPort(const std_msgs::Float64::ConstPtr &_msg);
+  void ros_callback_thrusterStrb(const std_msgs::Float64::ConstPtr &_msg);
   void ros_callback_menu(const std_msgs::String::ConstPtr &_msg);
 
   //======================================================//
